correção

+++ aulapratica7_lista7_codigo6_avaliativa.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 float soma(int n){
-    int soma=0;
+    float soma=0;
     for(int i = 1; i<=n; i++){
         soma += 1/i;
     }
